fix: switch to poetry to match main and update dockerfile
