+++ pyproject.toml
@@ -10,6 +10,16 @@ package-mode = false
 python = "^3.12"
 fastapi = "^0.110"
 uvicorn = "^0.29"
+debugpy = "^1.8.1"
+sqlalchemy = "^2.0.29"
+asyncpg = {extras = ["binary"], version = "^0.29.0"}
+alembic = "^1.13.1"
+pytest-asyncio = "~0.21"
+httpx = "^0.27.0"
+sqlmodel = "^0.0.16"
+black = "^24.4.2"
+ruff = "^0.4.2"
+pytest-postgresql = "^6.0.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8"
