--- requirements/base.txt
@@ -1,5 +1,0 @@
-sqlalchemy
-fastapi
-asyncpg
-uvicorn
-sqlmodel
