--- requirements/dev.txt
@@ -1,12 +1,0 @@
-sqlalchemy
-fastapi
-asyncpg
-psycopg[binary]
-uvicorn
-alembic
-pytest
-pytest_asyncio
-httpx
-sqlmodel
-black
-ruff
